1683_invalid_tweets.sql
This SQL query aims to identify invalid tweets from the Tweets table in a social media application. The table consists of two columns: tweet_id, which serves as the primary key, and content, which contains the text of the tweets.

A tweet is deemed invalid if the number of characters in its content exceeds 15. Therefore, the objective is to return the IDs of all such invalid tweets.